Move Indestructible into Metal, add Best Practice on statics
