+++ tt-train/sources/ttml/CMakeLists.txt
@@ -101,6 +101,7 @@ target_link_libraries(
         pthread
         atomic
         Metalium::Metal
+        Metalium::Metal::STL
         Metalium::TTNN
         Python::Python
         fmt::fmt-header-only
