+++ tt-train/sources/ttml/autograd/auto_context.hpp
@@ -4,10 +4,10 @@
 
 #pragma once
 
+#include <indestructible.hpp>
 #include <memory>
 #include <random>
 
-#include "core/indestructible.hpp"
 #include "core/mesh_device.hpp"
 #include "graph.hpp"
 
@@ -62,7 +62,7 @@ private:
     tt::tt_metal::distributed::MeshShape m_mesh_shape = {1, 1};
     std::unique_ptr<core::MeshDevice> m_device;
 
-    friend class core::Indestructible<AutoContext>;
+    friend class tt::stl::Indestructible<AutoContext>;
 };
 
 inline auto& ctx() {
